fix code to print info in dialog
